Apply some simplifications, most suggested by IntelliJ.
The most interesting thing I did manually was to migrate some of our internals off `Optional`: Now that we've been using nullness checking for a while, there's not much point.

RELNOTES=n/a
PiperOrigin-RevId: 761689235

+++ core/src/main/java/com/google/common/truth/MathUtil.java
@@ -65,7 +65,7 @@ final class MathUtil {
    * either {@code left} or {@code right} is infinite or NaN.
    */
   public static boolean equalWithinTolerance(float left, float right, float tolerance) {
-    return equalWithinTolerance((double) left, (double) right, (double) tolerance);
+    return equalWithinTolerance(left, right, (double) tolerance);
   }
 
   /**
@@ -87,6 +87,6 @@ final class MathUtil {
    * either {@code left} or {@code right} is infinite or NaN.
    */
   public static boolean notEqualWithinTolerance(float left, float right, float tolerance) {
-    return notEqualWithinTolerance((double) left, (double) right, (double) tolerance);
+    return notEqualWithinTolerance(left, right, (double) tolerance);
   }
 }
